IBPWF_code/requirements.txt
absl-py==0.9.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
gast==0.3.3
grpcio==1.27.2
h5py==2.10.0
importlab==0.5.1
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
mock==3.0.5
modAL==0.3.5
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
ninja==1.9.0.post1
notebook==6.0.3
numpy==1.18.1
pandas==0.24.2
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkg-resources==0.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.3
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytype==2020.5.13
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.9.1
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
skorch==0.7.0
tabulate==0.8.6
tensorboard==1.13.1
tensorboardX==2.0
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.2
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
torch==1.4.0
torchvision==0.5.0
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
typed-ast==1.4.1
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
zipp==1.2.0
